added simple routing mode and fixes for all tests still one minor error in testing but will be fixed on quote refactor updated server to handle and respond to errors as definined in mojaloop spec updated swagger to allow proper content-type
