+++ package.json
@@ -19,8 +19,8 @@
     "registry": "https://registry.npmjs.org"
   },
   "dependencies": {
+    "@hapi/boom": "7.4.2",
     "axios": "0.18.0",
-    "boom": "7.3.0",
     "good": "8.1.2",
     "good-console": "8.0.0",
     "good-squeeze": "5.1.0",
@@ -36,13 +36,16 @@
   "devDependencies": {
     "eslint": "5.16.0",
     "istanbul": "0.4.5",
+    "proxyquire": "^2.1.0",
+    "sinon": "7.3.2",
     "swagmock": "1.0.0",
-    "tape": "4.10.1"
+    "tape": "4.10.1",
+    "tapes": "4.1.0"
   },
   "scripts": {
     "start": "node src/server.js",
-    "test": "tape 'test/unit/**/*.js'",
-    "cover": "istanbul cover tape -- 'test/unit/**/*.js'",
+    "test": "tape 'test/unit/**/*.test.js'",
+    "cover": "istanbul cover tape -- 'test/unit/**/*.test.js'",
     "lint": "eslint .",
     "regenerate": "yo swaggerize:test --framework hapi --apiPath './src/interface/swagger.json'",
     "build": "docker build -t quoting-service:local -f ./Dockerfile ../",
