+++ src/handlers/bulkQuotes.js
@@ -32,7 +32,7 @@
 
 'use strict'
 
-const Boom = require('boom')
+const Boom = require('@hapi/boom')
 
 /**
  * Operations on /bulkQuotes
