+++ src/interface/swagger.json
@@ -12,9 +12,6 @@
   "schemes": [
     "http"
   ],
-  "consumes": [
-    "application/json"
-  ],
   "produces": [
     "application/json"
   ],
