+++ test/unit/bulkQuotes/{ID}/error.test.js
@@ -36,34 +36,59 @@ const Test = require('tape')
 const Hapi = require('hapi')
 const HapiOpenAPI = require('hapi-openapi')
 const Path = require('path')
-const Mockgen = require('../../../data/mockgen.js')
+const Mockgen = require('../../../util/mockgen.js')
+const helper = require('../../../util/helper')
 
 /**
- * Test for /quotes/{ID}/error
+ * Test for /bulkQuotes/{ID}/error
  */
-Test('/quotes/{ID}/error', function (t) {
+Test('/bulkQuotes/{ID}/error', function (t) {
   /**
-     * summary: QuotesByIDAndError
-     * description: If the server is unable to find or create a quote, or some other processing error occurs, the error callback PUT /quotes/&lt;ID&gt;/error is used. The &lt;ID&gt; in the URI should contain the quoteId that was used for the creation of the quote, or the &lt;ID&gt; that was used in the GET /quotes/&lt;ID&gt;.
+     * summary: BulkQuotesErrorByID
+     * description: If the server is unable to find or create a bulk quote, or another processing error occurs, the error callback PUT /bulkQuotes/&lt;ID&gt;/error is used. The &lt;ID&gt; in the URI should contain the bulkQuoteId that was used for the creation of the bulk quote, or the &lt;ID&gt; that was used in the GET /bulkQuotes/&lt;ID&gt;.
      * parameters: ID, body, Content-Length, Content-Type, Date, X-Forwarded-For, FSPIOP-Source, FSPIOP-Destination, FSPIOP-Encryption, FSPIOP-Signature, FSPIOP-URI, FSPIOP-HTTP-Method
      * produces: application/json
      * responses: 200, 400, 401, 403, 404, 405, 406, 501, 503
      */
-  t.test('test QuotesByIDAndError put operation', async function (t) {
+  t.test('test BulkQuotesErrorByID put operation', async function (t) {
     const server = new Hapi.Server()
 
     await server.register({
       plugin: HapiOpenAPI,
       options: {
-        api: Path.resolve(__dirname, '../../../src/interface/swagger.json'),
-        handlers: Path.join(__dirname, '../../../handlers'),
-        outputvalidation: true
+        api: Path.resolve(__dirname, '../../../../src/interface/swagger.json'),
+        handlers: Path.join(__dirname, '../../../../src/handlers'),
+        outputvalidation: false
       }
     })
 
+    await server.ext([
+      {
+        type: 'onPreResponse',
+        method: (request, h) => {
+          if (!request.response.isBoom) {
+          } else {
+            const error = request.response
+            error.message = {
+              errorInformation: {
+                errorCode: error.output.statusCode,
+                errorDescription: error.message,
+                extensionList:[{
+                  key: '',
+                  value: ''
+                }]
+              }
+            }
+            error.reformat()
+          }
+          return h.continue
+        }
+      }
+    ])
+
     const requests = new Promise((resolve, reject) => {
       Mockgen().requests({
-        path: '/quotes/{ID}/error',
+        path: '/bulkQuotes/{ID}/error',
         operation: 'put'
       }, function (error, mock) {
         return error ? reject(error) : resolve(mock)
@@ -78,7 +103,8 @@ Test('/quotes/{ID}/error', function (t) {
     // Mock request Path templates({}) are resolved using path parameters
     const options = {
       method: 'put',
-      url: '/fsp' + mock.request.path
+      url: mock.request.path,
+      headers: helper.defaultHeaders()
     }
     if (mock.request.body) {
       // Send the request body
@@ -87,17 +113,11 @@ Test('/quotes/{ID}/error', function (t) {
       // Send the request form data
       options.payload = mock.request.formData
       // Set the Content-Type as application/x-www-form-urlencoded
-      options.headers = options.headers || {}
-      options.headers['Content-Type'] = 'application/x-www-form-urlencoded'
+      options.headers = helper.defaultHeaders()
     }
-    // If headers are present, set the headers.
-    if (mock.request.headers && mock.request.headers.length > 0) {
-      options.headers = mock.request.headers
-    }
-
     const response = await server.inject(options)
 
-    t.equal(response.statusCode, 200, 'Ok response status')
+    t.equal(response.statusCode, 501, 'Not Implemented response status')
     t.end()
   })
 })
