+++ test/unit/quotes.test.js
@@ -36,35 +36,36 @@ const Test = require('tape')
 const Hapi = require('hapi')
 const HapiOpenAPI = require('hapi-openapi')
 const Path = require('path')
-const Mockgen = require('../../data/mockgen.js')
+const Mockgen = require('../util/mockgen.js')
+const helper = require('../util/helper')
 
 /**
- * Test for /quotes/{ID}
+ * Test for /quotes
  */
-Test('/quotes/{ID}', function (t) {
+Test('/quotes', function (t) {
   /**
-     * summary: QuotesByID
-     * description: The HTTP request GET /quotes/&lt;ID&gt; is used to get information regarding an earlier created or requested quote. The &lt;ID&gt; in the URI should contain the quoteId that was used for the creation of the quote.
-     * parameters: Accept
+     * summary: Quotes
+     * description: The HTTP request POST /quotes is used to request the creation of a quote for the provided financial transaction in the server.
+     * parameters: body, Accept, Content-Length, Content-Type, Date, X-Forwarded-For, FSPIOP-Source, FSPIOP-Destination, FSPIOP-Encryption, FSPIOP-Signature, FSPIOP-URI, FSPIOP-HTTP-Method
      * produces: application/json
      * responses: 202, 400, 401, 403, 404, 405, 406, 501, 503
      */
-  t.test('test QuotesByID get operation', async function (t) {
+  t.test('test Quotes post operation', async function (t) {
     const server = new Hapi.Server()
 
     await server.register({
       plugin: HapiOpenAPI,
       options: {
         api: Path.resolve(__dirname, '../../src/interface/swagger.json'),
-        handlers: Path.join(__dirname, '../../handlers'),
+        handlers: Path.join(__dirname, '../../src/handlers'),
         outputvalidation: true
       }
     })
 
     const requests = new Promise((resolve, reject) => {
       Mockgen().requests({
-        path: '/quotes/{ID}',
-        operation: 'get'
+        path: '/quotes',
+        operation: 'post'
       }, function (error, mock) {
         return error ? reject(error) : resolve(mock)
       })
@@ -77,8 +78,9 @@ Test('/quotes/{ID}', function (t) {
     // Get the resolved path from mock request
     // Mock request Path templates({}) are resolved using path parameters
     const options = {
-      method: 'get',
-      url: '/fsp' + mock.request.path
+      method: 'post',
+      url: mock.request.path,
+      headers: helper.defaultHeaders()
     }
     if (mock.request.body) {
       // Send the request body
@@ -87,75 +89,12 @@ Test('/quotes/{ID}', function (t) {
       // Send the request form data
       options.payload = mock.request.formData
       // Set the Content-Type as application/x-www-form-urlencoded
-      options.headers = options.headers || {}
-      options.headers['Content-Type'] = 'application/x-www-form-urlencoded'
-    }
-    // If headers are present, set the headers.
-    if (mock.request.headers && mock.request.headers.length > 0) {
-      options.headers = mock.request.headers
+      options.headers = helper.defaultHeaders()
     }
 
     const response = await server.inject(options)
 
     t.equal(response.statusCode, 202, 'Ok response status')
-    t.end()
-  })
-  /**
-     * summary: QuotesByID
-     * description: The callback PUT /quotes/&lt;ID&gt; is used to inform the client of a requested or created quote. The &lt;ID&gt; in the URI should contain the quoteId that was used for the creation of the quote, or the &lt;ID&gt; that was used in the GET /quotes/&lt;ID&gt;GET /quotes/&lt;ID&gt;.
-     * parameters: body, Content-Length
-     * produces: application/json
-     * responses: 200, 400, 401, 403, 404, 405, 406, 501, 503
-     */
-  t.test('test QuotesByID1 put operation', async function (t) {
-    const server = new Hapi.Server()
-
-    await server.register({
-      plugin: HapiOpenAPI,
-      options: {
-        api: Path.resolve(__dirname, '../../src/interface/swagger.json'),
-        handlers: Path.join(__dirname, '../../handlers'),
-        outputvalidation: true
-      }
-    })
-
-    const requests = new Promise((resolve, reject) => {
-      Mockgen().requests({
-        path: '/quotes/{ID}',
-        operation: 'put'
-      }, function (error, mock) {
-        return error ? reject(error) : resolve(mock)
-      })
-    })
-
-    const mock = await requests
-
-    t.ok(mock)
-    t.ok(mock.request)
-    // Get the resolved path from mock request
-    // Mock request Path templates({}) are resolved using path parameters
-    const options = {
-      method: 'put',
-      url: '/fsp' + mock.request.path
-    }
-    if (mock.request.body) {
-      // Send the request body
-      options.payload = mock.request.body
-    } else if (mock.request.formData) {
-      // Send the request form data
-      options.payload = mock.request.formData
-      // Set the Content-Type as application/x-www-form-urlencoded
-      options.headers = options.headers || {}
-      options.headers['Content-Type'] = 'application/x-www-form-urlencoded'
-    }
-    // If headers are present, set the headers.
-    if (mock.request.headers && mock.request.headers.length > 0) {
-      options.headers = mock.request.headers
-    }
-
-    const response = await server.inject(options)
-
-    t.equal(response.statusCode, 200, 'Ok response status')
     t.end()
   })
 })
