+++ test/unit/quotes/{ID}/error.test.js
@@ -32,39 +32,53 @@
 
 'use strict'
 
-const Test = require('tape')
+const Test = require('tapes')(require('tape'))
 const Hapi = require('hapi')
 const HapiOpenAPI = require('hapi-openapi')
 const Path = require('path')
-const Mockgen = require('../data/mockgen.js')
-
+const Mockgen = require('../../../util/mockgen.js')
+const helper = require('../../../util/helper')
+const Sinon = require('sinon')
+const Db = require('../../../../src/data/database')
 /**
- * Test for /bulkQuotes
+ * Test for /quotes/{ID}/error
  */
-Test('/bulkQuotes', function (t) {
+Test('/quotes/{ID}/error', t => {
+
+  let sandbox
+
+  t.beforeEach(test => {
+    sandbox = Sinon.createSandbox()
+    test.end()
+  })
+
+  t.beforeEach(test => {
+    sandbox.restore()
+    test.end()
+  })
   /**
-     * summary: BulkQuotes
-     * description: The HTTP request POST /bulkQuotes is used to request the creation of a bulk quote for the provided financial transactions in the server.
-     * parameters: body, Accept, Content-Length, Content-Type, Date, X-Forwarded-For, FSPIOP-Source, FSPIOP-Destination, FSPIOP-Encryption, FSPIOP-Signature, FSPIOP-URI, FSPIOP-HTTP-Method
+     * summary: QuotesByIDAndError
+     * description: If the server is unable to find or create a quote, or some other processing error occurs, the error callback PUT /quotes/&lt;ID&gt;/error is used. The &lt;ID&gt; in the URI should contain the quoteId that was used for the creation of the quote, or the &lt;ID&gt; that was used in the GET /quotes/&lt;ID&gt;.
+     * parameters: ID, body, Content-Length, Content-Type, Date, X-Forwarded-For, FSPIOP-Source, FSPIOP-Destination, FSPIOP-Encryption, FSPIOP-Signature, FSPIOP-URI, FSPIOP-HTTP-Method
      * produces: application/json
-     * responses: 202, 400, 401, 403, 404, 405, 406, 501, 503
+     * responses: 200, 400, 401, 403, 404, 405, 406, 501, 503
      */
-  t.test('test BulkQuotes post operation', async function (t) {
+  t.test('test QuotesByIDAndError put operation', async function (t) {
     const server = new Hapi.Server()
 
     await server.register({
       plugin: HapiOpenAPI,
       options: {
-        api: Path.resolve(__dirname, '../src/interface/swagger.json'),
-        handlers: Path.join(__dirname, '../handlers'),
+        api: Path.resolve(__dirname, '../../../../src/interface/swagger.json'),
+        handlers: Path.join(__dirname, '../../../../src/handlers'),
         outputvalidation: true
       }
     })
 
     const requests = new Promise((resolve, reject) => {
       Mockgen().requests({
-        path: '/bulkQuotes',
-        operation: 'post'
+        path: '/quotes/{ID}/error',
+        operation: 'put'
       }, function (error, mock) {
         return error ? reject(error) : resolve(mock)
       })
@@ -77,8 +91,9 @@ Test('/bulkQuotes', function (t) {
     // Get the resolved path from mock request
     // Mock request Path templates({}) are resolved using path parameters
     const options = {
-      method: 'post',
-      url: '/fsp' + mock.request.path
+      method: 'put',
+      url: mock.request.path,
+      headers: helper.defaultHeaders()
     }
     if (mock.request.body) {
       // Send the request body
@@ -87,17 +102,18 @@ Test('/bulkQuotes', function (t) {
       // Send the request form data
       options.payload = mock.request.formData
       // Set the Content-Type as application/x-www-form-urlencoded
-      options.headers = options.headers || {}
-      options.headers['Content-Type'] = 'application/x-www-form-urlencoded'
+      options.headers = helper.defaultHeaders()
     }
-    // If headers are present, set the headers.
-    if (mock.request.headers && mock.request.headers.length > 0) {
-      options.headers = mock.request.headers
-    }
+    sandbox.stub(Db.prototype, 'getParticipantEndpoint').callsFake(
+      (err, info) => {
+        console.info(info)
+      }
+    )
 
     const response = await server.inject(options)
 
-    t.equal(response.statusCode, 202, 'Ok response status')
+    t.equal(response.statusCode, 200, 'Ok response status')
     t.end()
   })
+  t.end()
 })
