+++ test/util/helper.js
@@ -1,4 +1,3 @@
-// (C)2018 ModusBox Inc.
 /*****
  License
  --------------
@@ -6,12 +5,6 @@
  The Mojaloop files are made available by the Bill & Melinda Gates Foundation under the Apache License, Version 2.0 (the "License") and you may not use these files except in compliance with the License. You may obtain a copy of the License at
  http://www.apache.org/licenses/LICENSE-2.0
  Unless required by applicable law or agreed to in writing, the Mojaloop files are distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
-
- Initial contribution
- --------------------
- The initial functionality and code base was donated by the Mowali project working in conjunction with MTN and Orange as service provides.
- * Project: Mowali
-
  Contributors
  --------------
  This is the official list of the Mojaloop project contributors for this file.
@@ -23,23 +16,38 @@
  their names indented and be marked with a '-'. Email address can be added
  optionally within square brackets <email>.
  * Gates Foundation
- - Name Surname <name.surname@gatesfoundation.com>
 
- * Henk Kodde <henk.kodde@modusbox.com>
- * Georgi Georgiev <georgi.georgiev@modusbox.com>
+ * Rajiv Mothilal <rajiv.mothilal@modusbox.com>
+
  --------------
  ******/
-
 'use strict'
-var Swagmock = require('swagmock')
-var Path = require('path')
-var apiPath = Path.resolve(__dirname, '../interface/swagger.json')
-var mockgen
 
-module.exports = function () {
-  /**
-     * Cached mock generator
-     */
-  mockgen = mockgen || Swagmock(apiPath)
-  return mockgen
+/**
+ * @function defaultHeaders
+ *
+ * @description This returns a set of default headers used for requests
+ *
+ * see https://nodejs.org/dist/latest-v10.x/docs/api/http.html#http_message_headers
+ *
+ * @returns {object} Returns the default headers
+ */
+function defaultHeaders() {
+  let destination = 'payeefsp'
+  let source = 'payerfsp'
+  let resource = 'quotes'
+  let version = '1.0'
+  // TODO: See API section 3.2.1; what should we do about X-Forwarded-For? Also, should we
+  // add/append to this field in all 'queueResponse' calls?
+  return {
+    'accept': `application/vnd.interoperability.${resource}+json;version=1`,
+    'fspiop-destination': destination ? destination : '',
+    'content-type': `application/vnd.interoperability.${resource}+json;version=${version}`,
+    'date': (new Date()).toUTCString(),
+    'fspiop-source': source
+  }
+}
+
+module.exports = {
+  defaultHeaders
 }
